move report tests into test_reports.py
